pyproject.toml
[project]
name = "yt-rss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.96",
    "feedgen>=1.0.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "yt-dlp>=2025.10.14",
]
